pyproject.toml
[project]
name = "ai-toolkit"
version = "0.1.0"
description = "Local-first AI CLI for building and managing agent and RAG projects"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project.scripts]
ai = "ai_cli.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]